oprava: vymazanie inzerata bez obrazka, refresh MyAdsProductAdapter.kt po vymazani inzeratu. pridana aktivita na zobrazenie produktu

+++ app/src/main/java/com/gensor/cyklobazar/database/FirestoreClass.kt
@@ -211,9 +211,11 @@ class FirestoreClass() : Database {
     metoda na vymazanie dokumentu z kolekcie podla id
      */
     override fun deleteProduct(product: Product) {
-        FirebaseStorage.getInstance()
-            .getReferenceFromUrl(product.image)
-            .delete()
+        if(product.image.isNotEmpty()){
+            FirebaseStorage.getInstance()
+                .getReferenceFromUrl(product.image)
+                .delete()
+        }
 
         when(product){
             is EBike -> {
